DEV: add skeleton for benchmarks
