+++ tests/test_process.cc
@@ -1,15 +1,15 @@
 #include "gtest/gtest.h"
 
-#include "process.h"
+#include "h5s3/private/process.h"
 
-using process = h5s3::testing::process;
+using process = h5s3::utils::process;
 
 TEST(process, clean_exit) {
     std::vector<std::string> args = {"python", "-c", "exit(0)"};
     process::environment env = {};
     process proc(args, env);
 
-    auto [code, state] = proc.join();
+    auto[code, state] = proc.join();
     EXPECT_EQ(code, 0);
     EXPECT_EQ(state, process::state::EXITED);
 }
@@ -19,7 +19,7 @@ TEST(process, error_exit) {
     process::environment env = {};
     process proc(args, env);
 
-    auto [code, state] = proc.join();
+    auto[code, state] = proc.join();
     EXPECT_EQ(code, 1);
     EXPECT_EQ(state, process::state::EXITED);
 }
@@ -31,7 +31,7 @@ TEST(process, environment) {
     process::environment env = {{"foo", "bar"}};
     process proc(args, env);
 
-    auto [code, state] = proc.join();
+    auto[code, state] = proc.join();
     EXPECT_EQ(code, 0);
     EXPECT_EQ(state, process::state::EXITED);
 }
@@ -44,7 +44,7 @@ TEST(process, interrupt) {
     EXPECT_TRUE(proc.running());
     proc.interrupt();
 
-    auto [code, state] = proc.join();
+    auto[code, state] = proc.join();
     EXPECT_EQ(code, SIGINT);
     EXPECT_EQ(state, process::state::SIGNALED);
 }
@@ -57,7 +57,7 @@ TEST(process, twice) {
     EXPECT_TRUE(proc.running());
     proc.interrupt();
     {
-        auto [code, state] = proc.join();
+        auto[code, state] = proc.join();
         EXPECT_EQ(code, SIGINT);
         EXPECT_EQ(state, process::state::SIGNALED);
         EXPECT_FALSE(proc.running());
@@ -76,7 +76,7 @@ TEST(process, terminate) {
     EXPECT_TRUE(proc.running());
     proc.terminate();
 
-    auto [code, state] = proc.join();
+    auto[code, state] = proc.join();
     EXPECT_EQ(code, SIGTERM);
     EXPECT_EQ(state, process::state::SIGNALED);
     EXPECT_FALSE(proc.running());
